Implementing a GDI-based texture acquisition system
This system is very inefficient and relies on the use of an intermediate system memory buffer, and furthermore will not yield the correct window texture in cases where advanced compositing is being used (Skype, for instance, looks particularly bad).

The ultimate aim will be to use the DWM APIs to achieve this effect, but this will entail a significant amount of reverse engineering and triage to get right.